+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 - Add an off-chain module for CTL projects.
 
+- Updated flake-parts, passing inputs now instead of self.
+
+- On-chain exports packages now, which can be accessed the normal way.
+
+- Basic hercules configuration support.
+
 ## 2.0.1 -- 2022-12-12
 
 - Vendored version of mkHackage instead of using now deprecated `haskell-nix-extra-hackage`.